add off-heap metrics add num active vertices add submissionclient akka actor add termination on unexpected container completion (YARN) add processing time propagation

+++ core/src/main/scala/io/joygraph/core/message/LoadingComplete.scala
@@ -1,3 +1,3 @@
 package io.joygraph.core.message
 
-case class LoadingComplete(workerId : Int, vertices : Int, edges : Int)
+case class LoadingComplete(workerId : Int, vertices : Long, edges : Long)
